Started to move share menu to bundle
